refactor: restructure writer module
